requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
google-genai>=1.0.0
boto3>=1.35.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
pillow>=10.0.0
requests>=2.31.0
gunicorn>=22.0.0
